refactor(ui): Se movio ui a 'engine/game' para futura refactorizacion.
